+++ hibernate/src/main/java/io/telicent/smart/cache/storage/hibernate/AbstractHibernateStorage.java
@@ -3,7 +3,10 @@
  */
 package io.telicent.smart.cache.storage.hibernate;
 
+import io.telicent.smart.cache.storage.hibernate.configuration.HibernateConfiguration;
+import io.telicent.smart.cache.storage.hibernate.configuration.JpaConfiguration;
 import jakarta.persistence.EntityManagerFactory;
+import jakarta.persistence.NamedQuery;
 import jakarta.persistence.Persistence;
 import jakarta.persistence.TypedQuery;
 import jakarta.persistence.criteria.CriteriaBuilder;
@@ -17,30 +20,44 @@ import java.util.function.Function;
 import java.util.function.Supplier;
 
 /**
- * A Hibernate based notifications store that can persist to some underlying database supported by Hibernate
+ * An abstract storage layer using Hibernate plus JPA to provide persistent {@link jakarta.persistence.Entity} centric
+ * storage.
  * <p>
- * This is the only persistent storage implementation currently provided, it is persistent provided that the underlying
- * database used is persistent.  For example in a test scenario you could use an H2 In-Memory store that would be
- * non-persistent, but in production we'd generally expect to be using a persistent Postgres database.
+ * Derived storage classes will typically extend this class and implement an appropriate storage interface for the
+ * service they are implementing storage for.  Derived classes have two main ways to interact with the underlying
+ * storage:
+ * </p>
+ * <ol>
+ *     <li>The {@link #begin()} method to start a new transaction.</li>
+ *     <li>
+ *     The various helper methods that take the {@link TransactionContext} returned by {@link #begin()} and use it
+ *     to carry out common operations e.g. {@link #getOrCreateByNaturalId(TransactionContext, Object, Class, Supplier)}
+ *     </li>
+ * </ol>
+ * <p>
+ * Please see the Javadoc on individual methods for more details of each methods functionality.
  * </p>
  */
 @ToString
-@SuppressWarnings("unused")
+@SuppressWarnings({"unused", "UnusedReturnValue"})
 public abstract class AbstractHibernateStorage implements AutoCloseable {
 
     @ToString.Exclude
     private final EntityManagerFactory entityManagerFactory;
 
     private volatile boolean closed = false;
+    /**
+     * The JDBC URL of the underlying data source
+     */
     protected final String jdbcUrl;
 
     /**
      * Creates a new hibernate backed store
      *
      * @param dbProperties    Database Connection properties, this should contain at least a
-     *                        {@value HibernateConfiguration#JAKARTA_PERSISTENCE_JDBC_URL} property to provide a JDBC
+     *                        {@value JpaConfiguration#JAKARTA_PERSISTENCE_JDBC_URL} property to provide a JDBC
      *                        connection to the database as well as any other relevant properties e.g.
-     *                        {@value HibernateConfiguration#JAKARTA_PERSISTENCE_JDBC_USER},
+     *                        {@value JpaConfiguration#JAKARTA_PERSISTENCE_JDBC_USER},
      *                        {@value HibernateConfiguration#HIBERNATE_DIALECT} etc.
      * @param persistenceUnit The name of the persistence unit to use, this should generally contain only the basic
      *                        configuration e.g. Entity Classes, JPA Provider, generic JPA configuration.  The actual
@@ -49,24 +66,20 @@ public abstract class AbstractHibernateStorage implements AutoCloseable {
      */
     public AbstractHibernateStorage(Properties dbProperties, String persistenceUnit) {
         this.entityManagerFactory = Persistence.createEntityManagerFactory(persistenceUnit, dbProperties);
-        this.jdbcUrl = dbProperties.getProperty(HibernateConfiguration.JAKARTA_PERSISTENCE_JDBC_URL);
+        this.jdbcUrl = dbProperties.getProperty(JpaConfiguration.JAKARTA_PERSISTENCE_JDBC_URL);
     }
 
     /**
      * Begins a transaction internally
-     * </p>
      * <p>
-     * Callers <strong>MUST</strong> ensure they use the return value in a try-with-resources block so that short-lived
-     * transactions are promptly commited/aborted as appropriate.
-     * </p>
-     * <p>
-     * If callers need to do multiple operations within a transaction they should use the various overloads of methods
-     * that take a {@link TransactionContext} as their first argument
+     * Callers <strong>MUST</strong> ensure they use the return value in a try-with-resources block so that transactions
+     * are promptly commited/rolled back as appropriate.  Callers should either {@link TransactionContext#commit()} the
+     * transaction to persist their changes, or {@link #close()} the transaction to roll back those changes.
      * </p>
      *
      * @return Transaction Context
      */
-    protected final TransactionContext beginInternal() {
+    protected final TransactionContext begin() {
         ensureNotClosed();
 
         // Start a new short-lived transaction
@@ -93,7 +106,7 @@ public abstract class AbstractHibernateStorage implements AutoCloseable {
      * @param entityClass Entity Class
      * @param creator     Supplier to create a new entity instance if the requested instance does not yet exist
      * @param <T>         Entity type
-     * @return Entity instance
+     * @return Entity instance, guaranteed to be non-null
      */
     protected <T> T getOrCreateByNaturalId(TransactionContext transaction, Object id, Class<T> entityClass,
                                            Supplier<T> creator) {
@@ -109,7 +122,19 @@ public abstract class AbstractHibernateStorage implements AutoCloseable {
     }
 
     /**
-     * Gets or creates an entity based on the outcome of a named query
+     * Gets or creates an entity based on the outcome of a named query.
+     * <p>
+     * If the named query returns a single result then the existing entity is returned, if it returns no results then
+     * the provided supplier is used to create the new entity instance and persist it.  Therefore, callers
+     * <strong>MUST</strong> ensure that they call {@link TransactionContext#commit()} on the provided transaction
+     * context otherwise the newly created entity won't be guaranteed to be persistent.
+     * </p>
+     * <p>
+     * The named query, combined with the given query parameters <strong>MUST</strong> yield either a single result, or
+     * no results, if it yields multiple results then an {@link IllegalStateException} is thrown.  This generally means
+     * that the named query you are referencing is not sufficiently selective to determine whether an entity already
+     * exists.
+     * </p>
      *
      * @param transaction     Transaction Context
      * @param entityClass     Entity Class
@@ -118,6 +143,7 @@ public abstract class AbstractHibernateStorage implements AutoCloseable {
      * @param creator         Supplier to create a new entity instance if the named query yields no result
      * @param <T>             Entity type
      * @return Entity instance
+     * @throws IllegalStateException Thrown if the named query returns more than one results
      */
     protected <T> T getOrCreateByNamedQuery(TransactionContext transaction, Class<T> entityClass, String queryName,
                                             Map<String, Object> queryParameters, Supplier<T> creator) {
@@ -128,6 +154,8 @@ public abstract class AbstractHibernateStorage implements AutoCloseable {
             T instance = creator.get();
             transaction.getEntityManager().persist(instance);
             return instance;
+        } else if (results.size() > 1) {
+            throw new IllegalStateException("Named Query " + queryName + " has more than one result");
         } else {
             return results.get(0);
         }
@@ -176,6 +204,12 @@ public abstract class AbstractHibernateStorage implements AutoCloseable {
      * The {@link jakarta.persistence.Entity} annotated class <strong>MUST</strong> have one, and only one, field
      * annotated with the Hibernate {@link org.hibernate.annotations.NaturalId} annotation.
      * </p>
+     * <p>
+     * As opposed to the {@link #loadByNaturalId(TransactionContext, Object, Class)} overload this method allows for a
+     * loader function that can mutate the loaded entity before returning it e.g. unwrapping a JPA
+     * {@link jakarta.persistence.Entity} into a plain POJO, calculating computed properties, forcing lazily loaded JPA
+     * properties to be fetched etc.
+     * </p>
      *
      * @param transaction Transaction Context
      * @param id          Natural ID
@@ -198,6 +232,15 @@ public abstract class AbstractHibernateStorage implements AutoCloseable {
 
     /**
      * Loads entities based on a JPA {@link jakarta.persistence.NamedQuery} defined on the entity class
+     * <p>
+     * The given entity class <strong>MUST</strong> have a JPA {@link jakarta.persistence.NamedQuery} annotation whose
+     * {@link NamedQuery#name()} matches the provided query name defined.
+     * </p>
+     * <p>
+     * This overload assumes that the named query needs no query parameters, if parameters are needed please use the
+     * {@link #loadByNamedQuery(TransactionContext, Class, String, Map)} or
+     * {@link #loadByNamedQuery(TransactionContext, Class, String, Map, Function)} overloads instead.
+     * </p>
      *
      * @param transaction Transaction Context
      * @param entityClass Entity Class
@@ -211,6 +254,17 @@ public abstract class AbstractHibernateStorage implements AutoCloseable {
 
     /**
      * Loads entities based on a JPA {@link jakarta.persistence.NamedQuery} defined on the entity class
+     * <p>
+     * The given entity class <strong>MUST</strong> have a JPA {@link jakarta.persistence.NamedQuery} annotation whose
+     * {@link NamedQuery#name()} matches the provided query name defined.
+     * </p>
+     * <p>
+     * The provided {@code queryParameters} map allows to pass in parameters for the named query, the values
+     * <strong>MUST</strong> match the expected JPA types e.g. if joining by entities then the caller
+     * <strong>MUST</strong> supply an instance of that entity, likely retrieved by a previous operation e.g.
+     * {@link #getOrCreateByNaturalId(TransactionContext, Object, Class, Supplier)} or
+     * {@link #loadByNaturalId(TransactionContext, Object, Class, Function)}.
+     * </p>
      *
      * @param transaction     Transaction Context
      * @param entityClass     Entity Class
@@ -226,6 +280,23 @@ public abstract class AbstractHibernateStorage implements AutoCloseable {
 
     /**
      * Loads entities based on a JPA {@link jakarta.persistence.NamedQuery} defined on the entity class
+     * <p>
+     * The given entity class <strong>MUST</strong> have a JPA {@link jakarta.persistence.NamedQuery} annotation whose
+     * {@link NamedQuery#name()} matches the provided query name defined.
+     * </p>
+     * <p>
+     * The provided {@code queryParameters} map allows to pass in parameters for the named query, the values
+     * <strong>MUST</strong> match the expected JPA types e.g. if joining by entities then the caller
+     * <strong>MUST</strong> supply an instance of that entity, likely retrieved by a previous operation e.g.
+     * {@link #getOrCreateByNaturalId(TransactionContext, Object, Class, Supplier)} or
+     * {@link #loadByNaturalId(TransactionContext, Object, Class, Function)}.
+     * </p>
+     * <p>
+     * As opposed to the {@link #loadByNamedQuery(TransactionContext, Class, String)} overload this method allows for a
+     * loader function that can mutate the loaded entities before returning it e.g. unwrapping a JPA
+     * {@link jakarta.persistence.Entity} into a plain POJO, calculating computed properties, forcing lazily loaded JPA
+     * properties to be fetched etc.
+     * </p>
      *
      * @param entityClass     Entity Class
      * @param queryName       Query Name
@@ -244,7 +315,10 @@ public abstract class AbstractHibernateStorage implements AutoCloseable {
     }
 
     /**
-     * Loads all entities
+     * Loads all entities of a given type
+     * <p>
+     * The provided entity class <strong>MUST</strong> be annotated with JPA {@link jakarta.persistence.Entity}
+     * </p>
      *
      * @param transaction Transaction Context
      * @param entityClass Entity Class
@@ -257,6 +331,14 @@ public abstract class AbstractHibernateStorage implements AutoCloseable {
 
     /**
      * Loads all entities
+     * <p>
+     * The provided entity class <strong>MUST</strong> be annotated with JPA {@link jakarta.persistence.Entity}
+     * </p>
+     * <p>
+     * As opposed to the {@link #loadAll(TransactionContext, Class)} overload this method allows for a loader function
+     * that can mutate the loaded entities before returning it e.g. unwrapping a JPA {@link jakarta.persistence.Entity}
+     * into a plain POJO, calculating computed properties, forcing lazily loaded JPA properties to be fetched etc.
+     * </p>
      *
      * @param transaction Transaction Context
      * @param entityClass Entity Class
@@ -283,6 +365,17 @@ public abstract class AbstractHibernateStorage implements AutoCloseable {
 
     /**
      * Prepares a named query
+     * <p>
+     * The given entity class <strong>MUST</strong> have a JPA {@link jakarta.persistence.NamedQuery} annotation whose
+     * {@link NamedQuery#name()} matches the provided query name defined.
+     * </p>
+     * <p>
+     * The provided {@code queryParameters} map allows to pass in parameters for the named query, the values
+     * <strong>MUST</strong> match the expected JPA types e.g. if joining by entities then the caller
+     * <strong>MUST</strong> supply an instance of that entity, likely retrieved by a previous operation e.g.
+     * {@link #getOrCreateByNaturalId(TransactionContext, Object, Class, Supplier)} or
+     * {@link #loadByNaturalId(TransactionContext, Object, Class, Function)}.
+     * </p>
      *
      * @param transaction     Transaction context
      * @param entityClass     Entity Class
@@ -293,7 +386,6 @@ public abstract class AbstractHibernateStorage implements AutoCloseable {
      */
     protected <T> TypedQuery<T> prepareNamedQuery(TransactionContext transaction, Class<T> entityClass,
                                                   String queryName, Map<String, Object> queryParameters) {
-        // Cache the named queries for performance reasons
         TypedQuery<T> query = transaction.getEntityManager().createNamedQuery(queryName, entityClass);
         if (MapUtils.isNotEmpty(queryParameters)) {
             for (Map.Entry<String, Object> entry : queryParameters.entrySet()) {