+++ hibernate/src/main/java/io/telicent/smart/cache/storage/hibernate/ShortLivedTransactionContext.java
@@ -7,15 +7,20 @@ import jakarta.persistence.EntityManager;
 import jakarta.persistence.EntityManagerFactory;
 import lombok.Getter;
 import org.hibernate.Session;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 import java.util.Objects;
 
 /**
  * Short-lived transaction Context, intended for use within a single method, or to be passed between several methods to
- * orchestrate a larger transaction.  Guarantees to commit/abort the transaction upon closure.
+ * orchestrate a larger transaction.  Guarantees to commit/abort the transaction upon closure, as well as closing the
+ * JPA {@link EntityManager} and Hibernate {@link Session}.
  */
 @Getter
 final class ShortLivedTransactionContext implements TransactionContext {
+
+    private static final Logger LOGGER = LoggerFactory.getLogger(ShortLivedTransactionContext.class);
 
     private final EntityManager entityManager;
     private final Session session;
@@ -54,10 +59,21 @@ final class ShortLivedTransactionContext implements TransactionContext {
     @Override
     public void close() {
         // If anything went wrong, and we're still in the transaction roll it back
-        if (this.entityManager.getTransaction().isActive()) {
-            this.entityManager.getTransaction().rollback();
+        // Do this in a try block as we want to guarantee to close the session and entity manager regardless
+        try {
+            if (this.isActive()) {
+                this.entityManager.getTransaction().rollback();
+            }
+        } catch (Throwable e) {
+            LOGGER.warn("Failed to rollback transaction", e);
         }
-        this.session.close();
-        this.entityManager.close();
+
+        // Make sure we make an attempt to close both the Hibernate Session and the JPA Entity Manager
+        // These are most likely tied to the same internal resources BUT we should make sure to close both regardless
+        try {
+            this.session.close();
+        } finally {
+            this.entityManager.close();
+        }
     }
 }