+++ hibernate/src/test/java/io/telicent/smart/cache/storage/hibernate/model/OrderManager.java
@@ -3,34 +3,37 @@
  */
 package io.telicent.smart.cache.storage.hibernate.model;
 
-import io.telicent.smart.cache.storage.hibernate.AbstractH2MemoryStorage;
+import io.telicent.smart.cache.storage.hibernate.AbstractHibernateStorage;
 import io.telicent.smart.cache.storage.hibernate.TransactionContext;
 
 import java.math.BigDecimal;
 import java.util.List;
 import java.util.Map;
+import java.util.Properties;
 
-public class OrderManager extends AbstractH2MemoryStorage {
+@SuppressWarnings({ "unused", "UnusedReturnValue" })
+public class OrderManager extends AbstractHibernateStorage {
     /**
      * Creates a new order manager
+     *
+     * @param dbProperties Database connection properties
      */
-    public OrderManager() {
-        super("hibernate-storage-example");
+    public OrderManager(Properties dbProperties) {
+        super(dbProperties, "hibernate-storage-example");
     }
 
     public List<Address> getAddresses() {
         ensureNotClosed();
-        try (TransactionContext transaction = this.beginInternal()) {
+        try (TransactionContext transaction = this.begin()) {
             return this.loadAll(transaction, Address.class);
         }
     }
 
     public List<Address> getAddresses(String postCode) {
         ensureNotClosed();
-        try (TransactionContext transaction = this.beginInternal()) {
-            List<Address> addresses =
-                    this.loadByNamedQuery(transaction, Address.class, "findByPostalCode",
-                                          Map.of("postalCode", postCode));
+        try (TransactionContext transaction = this.begin()) {
+            List<Address> addresses = this.loadByNamedQuery(transaction, Address.class, "findByPostalCode",
+                                                            Map.of("postalCode", postCode));
             transaction.commit();
             return addresses;
         }
@@ -38,19 +41,28 @@ public class OrderManager extends AbstractH2MemoryStorage {
 
     public Address saveAddress(String recipient, String nameOrNumber, String street, String city, String postalCode) {
         ensureNotClosed();
-        try (TransactionContext transaction = this.beginInternal()) {
+        try (TransactionContext transaction = this.begin()) {
             Address address = this.getOrCreateByNamedQuery(transaction, Address.class, "findByDetails",
                                                            Map.of("recipient", recipient, "nameOrNumber", nameOrNumber,
                                                                   "street", street, "city", city, "postalCode",
-                                                                  postalCode), () -> {
-                        Address newAddress = new Address();
-                        newAddress.setRecipient(recipient);
-                        newAddress.setNameOrNumber(nameOrNumber);
-                        newAddress.setStreet(street);
-                        newAddress.setCity(city);
-                        newAddress.setPostalCode(postalCode);
-                        return newAddress;
-                    });
+                                                                  postalCode),
+                                                           () -> new Address(null, recipient, nameOrNumber, street,
+                                                                             city, postalCode));
+            transaction.commit();
+            return address;
+        }
+    }
+
+    public Address badSaveAddress(String recipient, String nameOrNumber, String street, String city,
+                                  String postalCode) {
+        ensureNotClosed();
+        try (TransactionContext transaction = this.begin()) {
+            // This will always fail if the street has more than one address, this is included merely to allow for test
+            // validation and coverage of this corner case of behaviour.
+            Address address =
+                    this.getOrCreateByNamedQuery(transaction, Address.class, "findByStreet", Map.of("street", street),
+                                                 () -> new Address(null, recipient, nameOrNumber, street, city,
+                                                                   postalCode));
             transaction.commit();
             return address;
         }
@@ -58,21 +70,21 @@ public class OrderManager extends AbstractH2MemoryStorage {
 
     public List<Product> listProducts() {
         ensureNotClosed();
-        try (TransactionContext transaction = this.beginInternal()) {
+        try (TransactionContext transaction = this.begin()) {
             return this.loadAll(transaction, Product.class);
         }
     }
 
     public List<Product> listProducts(boolean inStock) {
         ensureNotClosed();
-        try (TransactionContext transaction = this.beginInternal()) {
+        try (TransactionContext transaction = this.begin()) {
             return this.loadByNamedQuery(transaction, Product.class, inStock ? "inStock" : "outOfStock");
         }
     }
 
     public Product getProduct(String code) {
         ensureNotClosed();
-        try (TransactionContext transaction = this.beginInternal()) {
+        try (TransactionContext transaction = this.begin()) {
             Product product = this.loadByNaturalId(transaction, code, Product.class);
             transaction.commit();
             return product;
@@ -81,17 +93,16 @@ public class OrderManager extends AbstractH2MemoryStorage {
 
     public void saveProduct(String code, String name, String description, BigDecimal price, long available) {
         ensureNotClosed();
-        try (TransactionContext transaction = this.beginInternal()) {
+        try (TransactionContext transaction = this.begin()) {
             this.getOrCreateByNaturalId(transaction, code, Product.class,
-                                        () -> new Product(null, code, name, description, price,
-                                                          available));
+                                        () -> new Product(null, code, name, description, price, available));
             transaction.commit();
         }
     }
 
     public void updateStock(String code, long diff) {
         ensureNotClosed();
-        try (TransactionContext transaction = this.beginInternal()) {
+        try (TransactionContext transaction = this.begin()) {
             Product product = this.loadByNaturalId(transaction, code, Product.class);
             if (product == null) {
                 throw new IllegalStateException("Product not found");